Eriytä aineistonäkymä omaksi komponentikseen ja käytä sitä kansalaisen nähtävilläolovaiheessa
