+++ src/pages/suunnitelma/[oid]/nahtavillaolo.tsx
@@ -1,4 +1,4 @@
-import React, { Key, ReactElement, useState } from "react";
+import React, { ReactElement, useState } from "react";
 import ProjektiJulkinenPageLayout from "@components/projekti/kansalaisnakyma/ProjektiJulkinenPageLayout";
 import Section from "@components/layout/Section";
 import KeyValueTable, { KeyValueData } from "@components/KeyValueTable";
@@ -7,26 +7,19 @@ import useTranslation from "next-translate/useTranslation";
 import { useProjektiJulkinen } from "src/hooks/useProjektiJulkinen";
 import { formatDate } from "src/util/dateUtils";
 import SectionContent from "@components/layout/SectionContent";
-import { Aineisto, Kieli, ProjektiTyyppi, Viranomainen } from "@services/api";
+import { Kieli, ProjektiTyyppi, Viranomainen } from "@services/api";
 import FormatDate from "@components/FormatDate";
 import JataPalautettaNappi from "@components/button/JataPalautettaNappi";
 import Notification, { NotificationType } from "@components/notification/Notification";
 import MuistutusLomakeDialogi from "@components/projekti/kansalaisnakyma/MuistutusLomakeDialogi";
 import useProjektiBreadcrumbsJulkinen from "src/hooks/useProjektiBreadcrumbsJulkinen";
 import Trans from "next-translate/Trans";
-import HassuAccordion from "@components/HassuAccordion";
-import { AineistoKategoria, aineistoKategoriat } from "common/aineistoKategoriat";
-import { Link, Stack } from "@mui/material";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import ButtonFlat from "@components/button/ButtonFlat";
+import KansalaisenAineistoNakyma from "@components/projekti/common/KansalaisenAineistoNakyma";
 
 export default function Nahtavillaolo({ setRouteLabels }: PageProps): ReactElement {
   const { t } = useTranslation("projekti");
   const { data: projekti } = useProjektiJulkinen();
   const kuulutus = projekti?.nahtavillaoloVaihe;
-
-  const [expandedAineistoKategoriat, setExpandedAineistoKategoriat] = useState<Key[]>([]);
-  const areAineistoKategoriesExpanded = !!expandedAineistoKategoriat.length;
 
   const velho = projekti?.velho;
   const [muistutusLomakeOpen, setMuistutusLomakeOpen] = useState(false);
@@ -60,17 +53,6 @@ export default function Nahtavillaolo({ setRouteLabels }: PageProps): ReactEleme
   const vastaavaViranomainen = velho.suunnittelustaVastaavaViranomainen
     ? velho.suunnittelustaVastaavaViranomainen
     : velho.tilaajaOrganisaatio;
-
-  const getAlaKategoryIds = (aineistoKategoriat: AineistoKategoria[]) => {
-    const keys: Key[] = [];
-    aineistoKategoriat.forEach((kategoria) => {
-      keys.push(kategoria.id);
-      if (kategoria.alaKategoriat) {
-        keys.push(...getAlaKategoryIds(kategoria.alaKategoriat));
-      }
-    });
-    return keys;
-  };
 
   return (
     <ProjektiJulkinenPageLayout selectedStep={2} title="Kuulutus suunnitelman nähtäville asettamisesta">
@@ -117,47 +99,7 @@ export default function Nahtavillaolo({ setRouteLabels }: PageProps): ReactEleme
             </SectionContent>
           </Notification>
         </SectionContent>
-        <h4 className="vayla-small-title">{t(`ui-otsikot.nahtavillaolo.esittelyaineisto_ja_suunnitelmat`)}</h4>
-        <SectionContent>
-          <Trans
-            i18nKey="projekti:info.nahtavillaolo.ei-rata.suunnitelmiin_on_mahdollista"
-            values={{
-              kuulutusVaihePaattyyPaiva: formatDate(kuulutus.kuulutusVaihePaattyyPaiva),
-            }}
-            components={{ p: <p />, b: <b /> }}
-          />
-          <ButtonFlat
-            type="button"
-            onClick={() => {
-              if (areAineistoKategoriesExpanded) {
-                setExpandedAineistoKategoriat([]);
-              } else {
-                setExpandedAineistoKategoriat(getAlaKategoryIds(aineistoKategoriat.listKategoriat()));
-              }
-            }}
-            iconComponent={
-              <span className="fa-layers">
-                <FontAwesomeIcon
-                  icon="chevron-down"
-                  transform={`down-6`}
-                  flip={areAineistoKategoriesExpanded ? "vertical" : undefined}
-                />
-                <FontAwesomeIcon
-                  icon="chevron-up"
-                  transform={`up-6`}
-                  flip={areAineistoKategoriesExpanded ? "vertical" : undefined}
-                />
-              </span>
-            }
-          >
-            {areAineistoKategoriesExpanded ? "Sulje" : "Avaa"} kaikki kategoriat
-          </ButtonFlat>
-          <AineistoKategoriaAccordion
-            aineistoKategoriat={aineistoKategoriat.listKategoriat()}
-            aineistot={kuulutus.aineistoNahtavilla}
-            expandedState={[expandedAineistoKategoriat, setExpandedAineistoKategoriat]}
-          />
-        </SectionContent>
+        <KansalaisenAineistoNakyma projekti={projekti} kuulutus={kuulutus} />
         <h4 className="vayla-small-title">{t(`ui-otsikot.nahtavillaolo.muistutuksen_jattaminen`)}</h4>
         <SectionContent>
           <JataPalautettaNappi
@@ -196,73 +138,3 @@ export default function Nahtavillaolo({ setRouteLabels }: PageProps): ReactEleme
     </ProjektiJulkinenPageLayout>
   );
 }
-
-interface AineistoKategoriaAccordionProps {
-  aineistoKategoriat?: AineistoKategoria[];
-  aineistot?: Aineisto[] | null;
-  expandedState: [React.Key[], React.Dispatch<React.Key[]>];
-}
-
-const AineistoKategoriaAccordion = (props: AineistoKategoriaAccordionProps) => {
-  const { t } = useTranslation("aineisto");
-  return props.aineistoKategoriat ? (
-    <HassuAccordion
-      items={props.aineistoKategoriat?.map((kategoria) => {
-        const aineistot = props.aineistot?.filter(
-          (aineisto) =>
-            kategoria.id === aineisto.kategoriaId ||
-            kategoria.alaKategoriat?.some((alakategoria) => alakategoria.id === aineisto.kategoriaId)
-        );
-        return {
-          title: `${t(`aineisto-kategoria-nimi.${kategoria.id}`)} (${aineistot?.length || 0})`,
-          content: (
-            <SuunnitelmaAineistoKategoriaContent
-              aineistot={aineistot}
-              kategoria={kategoria}
-              expandedState={props.expandedState}
-            />
-          ),
-          id: kategoria.id,
-        };
-      })}
-      expandedState={props.expandedState}
-    />
-  ) : null;
-};
-
-interface SuunnitelmaAineistoKategoriaContentProps {
-  aineistot?: Aineisto[];
-  kategoria: AineistoKategoria;
-  expandedState: [React.Key[], React.Dispatch<React.Key[]>];
-}
-
-const SuunnitelmaAineistoKategoriaContent = (props: SuunnitelmaAineistoKategoriaContentProps) => {
-  return (
-    <>
-      {!!props.aineistot?.length ? (
-        <Stack direction="column" rowGap={1.5}>
-          {props.aineistot
-            ?.filter((aineisto) => typeof aineisto.tiedosto === "string" && aineisto.kategoriaId === props.kategoria.id)
-            .map((aineisto) => (
-              <Stack direction="row" alignItems="flex-end" columnGap={2} key={aineisto.dokumenttiOid}>
-                <Link href={aineisto.tiedosto!} target="_blank" rel="noreferrer">
-                  {aineisto.nimi}
-                </Link>
-                <span>({"pdf" || aineisto.nimi.split(".").pop()})</span>
-                <a href={aineisto.tiedosto!} target="_blank" rel="noreferrer">
-                  <FontAwesomeIcon icon="external-link-alt" size="lg" className="text-primary-dark" />
-                </a>
-              </Stack>
-            ))}
-        </Stack>
-      ) : (
-        <p>Kategoriassa ei ole aineistoa</p>
-      )}
-      <AineistoKategoriaAccordion
-        aineistoKategoriat={props.kategoria.alaKategoriat}
-        aineistot={props.aineistot}
-        expandedState={props.expandedState}
-      />
-    </>
-  );
-};
